// apps/fusion/project/ws/app/src/lib/routes/infy/routes/knowledge-hub/models/knowledgeHub.model.ts
export interface IAuthors {
  id: string
  mailId: string
  name: string
}
export interface ISearchObj {
  searchQuery: string
  from: number
  size: number
}
export interface ISearchObjForSearch {
  searchQuery: string
  from: number
  size: number
  category: string
  filter: string
}
export interface ISearchObjForView {
  category: string
  itemId: string
  source: string
}
export interface IProjectRisks {
  description: string
  exposure: string
  id: string
  name: string
}
export interface IProjectContribution {
  itemId: number
  title: string
  url: string
}
// ...
export interface IKhubPMDM {
  id: string
  mailId: string
  name: string
}
export interface IKhubProjectTools {
  actualHoursSaving: string
  indicativeEffortSaving: string
  lcStage: string
  name: string
  projectedEffortSaving: string
  purpose: string
  type: string
}
export interface IKhubProStrategies {
  description: string
  id: string
  name: string
}
export interface IKhubProReusable {
  id: string
  name: string | null
}
export interface IItemsList {
  data: (IKhubProStrategies | IKhubProjectTools | IProjectRisks | IProjectContribution)[]
  type: string
  itemsMinShown: number
  headingText: string
}
export interface IKhubProject {
  mstProjectCode: string
  dateStartDate: string
  dateEndDate: string
  mstLocation: string
  pu: string
  puName: string
  ibu: string
  ibuName: string
  du: string
  status: string
  mstProjectName: string
  mstDomain: string
  mstServiceOffering: string
  mstContractType: string
  mstProcessType: string
  mstProjectScope: string
  mstBusinessContext: string
  mstInfosysRole: string
  mstInfyObjectives: string
  projectClassIndicator: string
  dateLastModified: string
  dateCreated: string
  itemId: number
  serviceLine: string
  projecttype: string | null
  primaryIndustry: string | null
  secondayIndustry: string | null
  tertiaryIndustry: string | null
  requestType: string | null
  riskLevel: string | null
  typeOfApplicationMaintained: string | null
  technology: string | null
  natureOfWork: string | null
  packageOrToolUsed: string | null
  modeOfTesting: string | null
  typeOfTesting: string | null
  isAgile: string | null
  sizingTechnique: string | null
  projectLCStage: string | null
  isAccessRestricted: string
  extractedText: string
  qy: string[]
  strategies: IKhubProStrategies[]
  contributions: IProjectContribution[]
  tools: IKhubProjectTools[]
  reuseComponents: IKhubProReusable[]
  risks: IProjectRisks[]
  topics: string[]
  pm: IKhubPMDM[]
  dm: IKhubPMDM[]
}
// ...
export interface IKhubItemTile {
  source?: string
  url?: string
  author?: IAuthors[]
  itemType?: string
  description?: string
  topics?: string[]
  category?: string
  itemId?: number
  noOfViews?: number
  title?: string
  restricted?: string
  dateCreated?: Date
  projectScope?: string
  businessContext?: string
  color?: string
  dm?: IAuthors[]
  pm?: IAuthors[]
  objectives?: string
  risks?: IProjectRisks[]
  sourceId?: number
  contribution?: IProjectContribution[]
}